requirements.txt
absl-py==1.4.0
asgiref==3.6.0
attrs==23.1.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.18
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Flask==2.2.5
Flask-Cors==4.0.0
flask-requests==0.0.14
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
fonttools==4.38.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
logger==1.4
MarkupSafe==2.1.2
matplotlib==3.5.3
mediapipe==0.9.0.1
numpy==1.21.6
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
pampy==0.3.0
pandas==1.1.5
Pillow==9.5.0
protobuf==3.20.3
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
SQLAlchemy==2.0.21
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
absl-py==1.4.0
asgiref==3.6.0
attrs==23.1.0
blinker==1.6.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
cycler==0.11.0
defusedxml==0.7.1
Django==3.2.18
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Flask==2.2.5
Flask-Cors==4.0.0
flask-requests==0.0.14
Flask-SQLAlchemy==3.0.5
flatbuffers==23.5.26
fonttools==4.38.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
logger==1.4
MarkupSafe==2.1.2
matplotlib==3.5.3
mediapipe==0.9.0.1
numpy==1.21.6
oauthlib==3.2.2
opencv-contrib-python==4.8.0.76
opencv-python==4.8.0.76
packaging==23.1
pampy==0.3.0
pandas==1.1.5
Pillow==9.5.0
protobuf==3.20.3
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
SQLAlchemy==2.0.21
sqlparse==0.4.3
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
Werkzeug==2.2.3
zipp==3.15.0
